+++ Environments/set_plotting_env.sh
