--- Generation_Data/datasize.txt
@@ -1,1 +1,0 @@
-0.264925;
