--- Generation_Data/highfive.txt
@@ -1,1 +1,0 @@
-1
